ProductEdit async added
Beware an bug was intruduced with this comit, that contains problem when 2 users edit the same product, the image list in edit view is populated with the same image, even though there was only one selected on the product creation.

it has something to do with EF db.ProductImages retrival process, since it only retrives one entity not all. or something like that in the product->viewModel mapping method.